Refactored Translatable to populate all languages
Originally, content in a secondary language would fall through
to the default language. Not providing a translation would show
this content in the default language.
The fallback mechanism turned out to be rather straining on the database.

We've updated the behavior to solve the problem
at write-time, rather than at read-time.  If translations are not provided, we
will populate those records with a copy of the original default language content.