Include `lang` in unique key for i18n slugs
This avoids having `abc` as slug in the primary language and `abc-2` in
the secondary language.
For I18n models, the spawner includes `lang` in the unique key
constraint to make this happen.

Sluggable is also refactored to reflect this change. It will use the
given language when checking the slug's uniqueness.

Last but not least: a bug popped up after the recent Translatable
refactoring where an exception would be thrown when explicitly saving a
certain slug as opposed to auto-generating one thru the Sluggable
behavior. This feature fixes that.