Added monolog to dependencies
We're not actually using it, but this is used to lock the version to one
a bit lower than `dev-master`.
Since Sentry is requiring monolog: "*" it always takes the latest, which
requires PHP 7.
Since we have several installations even on 5.3, we have to prevent the
higher Monolog version from being installed.