Refactored spawner to make hasOne columns editable
Columns like `author_id` were previously hidden and not editable. This
made sense in the context of the old CMS, but is weird when considering
the new CMS. This has been refactored to better refect the nature of
these columns.

Additionally a `model` property is given to these column in order to
programmatically figure out the other end of the relationship, e.g.
linking `author_id` to the model `User`.