Fixed RelationManager::isInvalidReferenceException
It would report TRUE for any exception because it checked the length of
sscanf(), but this will always report the number of variables in a
string, even if they can't be resolved. (e.g. you end up with an array
containing NULLs that'll still count() up to the number you'd expect)