Read Sentry API URL from .env before constant
Right now, `index.php` is the sole entrypoint which configures Sentry,
but it's not the only entrypoint of a Garp application.
CLI commands actually use `garp/scripts/garp.php`. Both of these use
`init.php` to configure a bunch of stuff, so that should be the place
where things like Sentry are configured.

Thus: this refactors `init.php` so that it tries to get Sentry's API
URL first from `getenv()`, allowing users to maintain Sentry's API URL
from a `.env` file.
To maintain backward compatibility it will still check for the existence
of the constant as before.