Resolve BASE_PATH using cwd on command-line
This fixes an issue where the BASE_PATH is incorrectly determined in
case Garp is used as symlink to develop locally.
In those cases the location of the garp.php script is disconnected from
the project-root (since it's a symlink) and the only way to figure from
which project garp.php was called is from the CLI's current
working directory.