Deprecate Semver in favor of VERSION
A simpler VERSION file can be maintained based on `git describe`. It no
longer needs to be in version control, decoupling version from the
codebase, which is a good thing.

This makes sure Garp can handle the absence of a `.semver` file. It will
use a `VERSION` file whenever a version is required, but note that this
is an optional file. Most code should not rely on the presence of a
certain version string in the code.