Check for interactivity before reading STDIN
posix_isatty can be used to determine wether data is being piped into a
script.

Using this, we can revert stream_set_blocking() to true, in order to
make the script wait for all input.

Specifically fixes a case where remote data (which takes a while to be
downloaded) from 12g was not being picked up by garp.php.