Remove the DefaultSortable behavior
This behavior has proven to be error-prone, quick to break and
inflexible. It just does not offer enough worth to remain
in existence and is hereby terminated.

A breaking change, which should increase Garp version to `3.17.0`, since
queries might yield different results. However, I'm confident
queries in production are already outfitted with custom
`ORDER` clauses in cases where order matters.