Make Garp_Db_Table_Rowset return instances of static to make it extendable
Sometimes we want to extend the rowset to allow some logic on a collection of rows. Because the filter functions etc. used to return instances of the self, the return rowsets would be of Garp_Db_Table_Rowset instead of whatever type of rowset the function was called. This is now fixed by returning instances of static.