Replace path repository by `replace`
The path repository works for a checkout of Garp, but not when loaded as
dependency of an application.

By adding phpoffice-phpexcel to the classmap and removing it as a
dependency the package becomes part of Garp.

Remove PCLZip fixes a conflict with phpoffice/spreadsheet because that
package requires the package pclzip/pclzip which contains the same
logic. This means that creating a zipped spreadsheet is not possible
anymore with Garp, but it's not currently used by Garp. So no problem to
remove it.